+++ auralogger/utils/__init__.py
@@ -1,7 +1,6 @@
 """Utility exports for host integration and testing helpers."""
 
 from auralogger.utils.backend_origin import (
-    build_create_browser_logs_url,
     build_proj_auth_url,
     build_project_logs_url,
     resolve_api_base_url,
@@ -13,7 +12,6 @@ from auralogger.utils.env_config import (
     ENV_PROJECT_SESSION,
     ENV_PROJECT_STYLES,
     ENV_PROJECT_TOKEN,
-    ENV_PROJECT_SECRET,
     ENV_USER_SECRET,
     ENV_VITE_PROJECT_TOKEN,
 )
@@ -22,7 +20,6 @@ from auralogger.utils.http_utils import parse_error_body
 __all__ = [
     "build_proj_auth_url",
     "build_project_logs_url",
-    "build_create_browser_logs_url",
     "resolve_api_base_url",
     "resolve_ws_base_url",
     "parse_error_body",
@@ -30,7 +27,6 @@ __all__ = [
     "ENV_NEXT_PUBLIC_PROJECT_TOKEN",
     "ENV_VITE_PROJECT_TOKEN",
     "ENV_USER_SECRET",
-    "ENV_PROJECT_SECRET",
     "ENV_PROJECT_ID",
     "ENV_PROJECT_SESSION",
     "ENV_PROJECT_STYLES",